Remove duplicated rows in genes.tsv and orthologs file.
Some species seem to return duplicated entries for a few genes. For example,
for the rat the gene ENSRNOG00000032677/Tssc4 is returned twice - investigating
further, the two entries have different chromosomal bands. I don't know if
this is a bug; looking in the genome browser, only one of the bands appears
to be correct.